requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
attrs==22.1.0
Automat==22.10.0
boto3==1.24.42
botocore==1.27.96
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
cssselect==1.2.0
defusedxml==0.7.1
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-recaptcha==3.0.0
django-rest-registration==0.7.2
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.22.1
filelock==3.8.0
google-auth==2.14.1
google-auth-oauthlib==0.7.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gspread==5.4.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itemadapter==0.7.0
itemloaders==1.0.6
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.0
lxml==4.9.1
MarkupSafe==2.1.1
mysql==0.0.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==21.3
parsel==1.7.0
Pillow==9.0.1
promise==2.3
Protego==0.2.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
PyJWT==2.6.0
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
Rx==1.6.1
s3transfer==0.6.0
Scrapy==2.7.1
scrapy-djangoitem==1.1.1
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
text-unidecode==1.3
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
w3lib==2.0.1
zope.interface==5.5.2
